+++ generate_id.js
@@ -1,0 +1,3 @@
+export default function generateId() {
+    return (new Date().getTime() / 1000 | 0).toString(16) + 'xxxxxxxxxxxxxxxx'.replace(/[x]/g, () => (Math.random() * 16 | 0).toString(16)).toLowerCase();
+}